log-generator/requirements.txt
faker==24.9.0
python-dotenv==1.0.1
pytest==8.2.0
